STYLE: Remove `static_cast<unsigned int>` from dimension constants
Dimensions are already `unsigned int`, so these casts had no effect anyway.

Replaced `static_cast<unsigned int>\(([A-Za-z]+Dimension)\)` with `$1`, using
regular expressions.

Replaced `static_cast<unsigned int>(Self::CellDimension)` with `CellDimension`.

+++ Modules/Core/Common/include/itkInPlaceImageFilter.hxx
@@ -57,10 +57,9 @@ InPlaceImageFilter<TInputImage, TOutputImage>::InternalAllocateOutputs()
   // additionally the buffered and requested regions of the input and
   // output must match.
   bool rMatch = true;
-  if (inputPtr != nullptr &&
-      static_cast<unsigned int>(InputImageDimension) == static_cast<unsigned int>(OutputImageDimension))
+  if (inputPtr != nullptr && InputImageDimension == OutputImageDimension)
   {
-    for (unsigned int i = 0; i < static_cast<unsigned int>(InputImageDimension); ++i)
+    for (unsigned int i = 0; i < InputImageDimension; ++i)
     {
       if (inputPtr->GetBufferedRegion().GetIndex(i) != outputPtr->GetRequestedRegion().GetIndex(i))
       {
